admin/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
pika==1.2.1
SQLAlchemy==1.4.39
mysqlclient==2.1.0
python-dotenv==0.20.0